Increase the RANGE() precision due to rounding issues.
git-svn-id: svn://svn.berlios.de/gale/trunk@1403 2d798ef8-6e06-0410-941a-9092aa770104

+++ gale2/include/gale/math/color.h
@@ -59,8 +59,8 @@ struct ColorChannel
     }
 
     /// Returns the value range allowed for a color channel.
-    static float RANGE() {
-        return static_cast<float>(MAX_VALUE())-MIN_VALUE();
+    static double RANGE() {
+        return static_cast<double>(MAX_VALUE())-MIN_VALUE();
     }
 };
